chore: move FuzzAttribute, EDistributionType & FuzzParameter to Confirma.Fuzz namespace
